+++ lib/helpers/show_number_picker.dart
@@ -4,13 +4,15 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_material_pickers/dialogs/scroll_picker_dialog.dart';
 
+import '../flutter_material_pickers.dart';
+
 /// Allows selection of a number via a slot machine carousel
-void showMaterialNumberPicker({
+Future<int?> showMaterialNumberPicker({
   required BuildContext context,
   String? title,
   required final int minNumber,
   required final int maxNumber,
-  required final int selectedNumber,
+  final int? selectedNumber,
   final int step = 1,
   Color? headerColor,
   Color? headerTextColor,
@@ -23,20 +25,21 @@ void showMaterialNumberPicker({
   ValueChanged<int>? onChanged,
   VoidCallback? onConfirmed,
   VoidCallback? onCancelled,
+  Transformer<int>? transformer,
 }) {
-  List<String> items = [];
+  List<int> items = [];
 
   for (int i = minNumber; i <= maxNumber; i += step) {
-    items.add(i.toString());
+    items.add(i);
   }
 
-  showDialog<String>(
+  return showDialog<int>(
     context: context,
     builder: (BuildContext context) {
-      return ScrollPickerDialog(
+      return ScrollPickerDialog<int>(
         items: items,
         title: title,
-        initialItem: selectedNumber.toString(),
+        selectedItem: selectedNumber,
         headerColor: headerColor,
         headerTextColor: headerTextColor,
         backgroundColor: backgroundColor,
@@ -48,8 +51,12 @@ void showMaterialNumberPicker({
       );
     },
   ).then((selection) {
-    if (onChanged != null && selection != null) onChanged(int.parse(selection));
-    if (onCancelled != null && selection == null) onCancelled();
-    if (onConfirmed != null && selection != null) onConfirmed();
+    if (selection != null) {
+      onChanged?.call(selection);
+      onConfirmed?.call();
+    } else {
+      onCancelled?.call();
+    }
+    return selection;
   });
 }